move thread draft logic to useNewThreadForm hook
